Allow multiple JRuby contexts in one JVM
zweikopf.multi ns has an API similar to zweikopf.core, but all functions
take an extra argument: the JRuby context in which the operation should be
executed. #'zweikopf.multi/context allows to create such a context. This
enables having multiple independent JRuby contexts in a single JVM.

zweikopf.core becomes a thin layer using the newly introduced
zweikopf.multi with the JRuby context set to a singleton context.
Neither the behaviour nor the API of this namespace is modified in this
patch.